+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.6-beta
+* Fixed issue with qualitykiosk app desired capability
+* Moved QualityKiosk capability mapper to QualityKioskDriver
+* Added support when `app` capability is not passed
+
 # 0.1.5-beta
 * Support for PCloudy
 * Support for QualityKiosk